move `{key_image,output}` handling `add_block()` -> `add_tx()`
